big changes -> request and community are the same dynamic page
