+++ src/utils/routes.ts
@@ -2,10 +2,10 @@ export type Routes = string[];
 
 export const publicRoutes: Routes = ['/', '/stories', '/proposals'];
 
-export const privateRoutes: Routes = ['/profile', '/settings', '/requests', `/requests/[id]`];
+export const privateRoutes: Routes = ['/profile', '/settings'];
 
 export const demoRoutes: Routes = [];
 
-export const beneficiaryRoutes: Routes = ['/beneficiary', '/requests', `/requests/[id]`];
+export const beneficiaryRoutes: Routes = ['/beneficiary', '/requests', '/communities', '/communities/[id]'];
 
-export const managerRoutes: Routes = [];
+export const managerRoutes: Routes = ['/communities', '/communities/[id]'];
